Fix Constant conversions, part 10

+++ main1048.cpp
@@ -38,7 +38,7 @@ int main()
  Xf1c = (short)-320743561;
  tAc2 = (short)-94334232;
  lU_5 = (short)634134418;
- oELN01T = -2000660925;
+ oELN01T = (short)-2000660925;
  nhU_X = 1347239331;
  r2Rh_ = (short)1433981282;
  mnJgBfk_3 = (short)993498213;
@@ -48,7 +48,7 @@ int main()
  E3oqe4Ik = (short)-1099038588;
  l4Fy533Emd = (short)-1288244192;
  uRG5Vf = (short)-1955974764;
- BTxsZ = -934194130;
+ BTxsZ = (short)-934194130;
  C5tG9Y = -434220620;
  ykprBE = (short)1018268172;
  PSgs = (short)671593453;
@@ -58,7 +58,7 @@ int main()
  vn12Kz = (short)-1234771783;
  K_RN_D81P = (short)-575084118;
  V8E1iC__x6 = (short)-1784241334;
- wya7sM_ = 1178054150;
+ wya7sM_ = (short)1178054150;
  s0tOFRY7IK = 1713230019;
  F3j1y6 = (short)-1775403824;
  t5OT_L = (short)-1861851719;
@@ -68,7 +68,7 @@ int main()
  zGJR = (short)1316086969;
  E00aE3 = (short)-782457207;
  a58zGY2S2 = (short)-806523845;
- Kl8R_1vcaC = -585839864;
+ Kl8R_1vcaC = (short)-585839864;
  O4ux = -42478139;
  Bm4_56U = (short)2001278543;
  M15h = (short)-1794672295;
