+++ main1135.cpp
@@ -24,7 +24,7 @@ int main(void)
   DaCTX17r  =  (short)970311864;
   M6XvJoln  =  (short)1340003466;
   cKe9ds3h  =  (short)1688513628;
-  Ts5uI912C  =  2080079096;
+  Ts5uI912C  =  (short)2080079096;
   GuH42MEB  =  742334245;
   jT45uZJsu  =  (short)1703341761;
   Ubf1XfMkW  =  (short)471611830;
@@ -34,7 +34,7 @@ int main(void)
   hnW1CA  =  (short)-117980695;
   IKEkjLLtz  =  (short)-561920606;
   ONkC  =  (short)81097547;
-  LIi1Re2vFe  =  1288599581;
+  LIi1Re2vFe  =  (short)1288599581;
   m_F9Gwx  =  -1012763619;
   Rbqf7pE1  =  (short)503326375;
   Ag9iI1S  =  (short)998923591;
@@ -44,7 +44,7 @@ int main(void)
   u7968zi02  =  (short)31952100;
   Nn5xam  =  (short)1088426915;
   a7ni  =  (short)81916014;
-  pJ9Z62yaHJ  =  1338748173;
+  pJ9Z62yaHJ  =  (short)1338748173;
   bocer4  =  483817355;
   I6s406n  =  (short)714781570;
   RQ55  =  (short)391007324;
@@ -54,7 +54,7 @@ int main(void)
   Ip43hu  =  (short)25313957;
   tDc50  =  (short)43440187;
   A9vXmR  =  (short)-599678289;
-  PHsCRuUO2  =  -654827627;
+  PHsCRuUO2  =  (short)-654827627;
   hEes2  =  241199084;
   O_Kl  =  (short)684153409;
   v_2hB  =  (short)-571534512;
@@ -64,7 +64,7 @@ int main(void)
   E4Xy_  =  (short)-566838952;
   Ti3_  =  (short)608970325;
   mhU7W7  =  (short)1643216997;
-  sizjkX3_T  =  -1585413173;
+  sizjkX3_T  =  (short)-1585413173;
   rll5  =  1744163886;
   w_KM2M02Tv  =  (short)1411161794;
   UQw9vw  =  (short)-893102532;
